requirements.txt
Flask==3.0.0
google-cloud-storage==2.13.0
google-cloud-firestore==2.14.0
gunicorn==21.2.0
Pillow==10.1.0